ajout dun static cast dans la verification du client max body size, sinon ca compile pas

+++ srcs/Request.cpp
@@ -6,7 +6,7 @@
 /*   By: jpiech <jpiech@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/23 14:16:19 by qsomarri          #+#    #+#             */
-/*   Updated: 2025/09/04 12:05:36 by jpiech           ###   ########.fr       */
+/*   Updated: 2025/09/04 12:09:35 by jpiech           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -187,7 +187,7 @@ void Request::parsBody()
 	{
 		this->_body_len = std::atoi(this->_headers["CONTENT-LENGTH"].c_str());
 		if (this->_config.find("client_max_body_size") != this->_config.end() && !this->_config["client_max_body_size"].empty() && this->_config["client_max_body_size"] != "0")
-			if (this->_body_len > std::atoi(this->_config["client_max_body_size"].c_str()))
+			if (this->_body_len > static_cast<size_t>(atoi(this->_config["client_max_body_size"].c_str())))
 				return ((void)setStatus("413"));
 		// std::cout << "msg.size()= " << msg.size() << std::endl;
 		// std::cout << "body_len= " << this->_body_len << std::endl;
